refactor: split agent projection families
